fix parameter problem
Platform parameters are now handled via minimist. When parsing parameters
commander wants to handle the whole thing. That didn't work because plugin
needs to perform stuff on the commander object before running anything.
To load properly plugins, the cozy light home parameters was required (need to
know where is the config and where plugins are stored).

So parameters parsing was required anyway. That's why I use here another lib to
perform it properly.